==> scripts/02_get_municipios_selenium.ipynb <==
<br>

# Método POST

Notei que o site usa uma API "por atrás" para retornar as informações.


import time

from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import Select

from sp_tjsp_divadmin.my_driver import Driver
from sp_tjsp_divadmin.my_paths import adds_path, driver_path, logs_path

<br>

----

## Driver

Inicialmente, na primeira abordagem, tentei acessar o site da Lista Telefônica para obter os dados do TJSP.


driver = Driver(
    my_driver_path=driver_path,
    my_logs_path=logs_path,
    verify_ssl=False,
)
driver.add_extension_xpath(my_adds_path=adds_path)

# Go TJSP
URL = 'https://www.tjsp.jus.br/ListaTelefonica'
driver.get(URL)

# Parâmetros
procurar_por = 'Município'
text_procurar = 'Santos'

# Tipo de Busca
select = Select(driver.find_element(By.XPATH, '//*[@id="btnTipoBusca"]'))

# Opções para Procura
option_list = select.options
option_list = [x.text for x in option_list]

# sss
if procurar_por in option_list:
    select.select_by_visible_text(procurar_por)
else:
    raise Exception(f'Valor deve estar entre {option_list}')

<br>

Dai notei que a cada letra digitada, uma lista de municípios apareciam, indicando que algum método POST estava atuando.

Optei por ver o método POST e trabalhar com ele, mudando a abordagem!


field_xpath = driver.find_element(By.XPATH, '//*[@id="Descricao"]')
field_xpath.clear()
field_xpath.send_keys(Keys.BACKSPACE * 50)

for letra in text_procurar:
    print(letra)
    time.sleep(1)
    field_xpath.send_keys(letra)

driver.quit()
