==> scripts/06_geodata.ipynb <==
<br>

# Introdução

Juntar Código da CJ na tabela


import pandas as pd
from open_geodata import converts, geo
from sp_tjsp_divadmin.my_paths import (
    output_path_geo,
    output_path_gpkg,
    output_path_tab,
)

<br>

## Dados Espaciais


geo.get_dataset_names()

# Pega Delimitação dos Municípios
gdf_mun = geo.load_dataset('geo.sp.sp_250k_wgs84')

# Delete Columns
gdf_mun = gdf_mun.drop(['municipio_nome'], axis=1, errors='ignore')
gdf_mun.head(2)

# Pega os Nomes Corretos
df_geo_mun = geo.load_dataset('tab.sp.tab_municipio_nome')
df_geo_mun.head(2)

# Junta Delimitação com Nomes Corretos
gdf_mun = gdf_mun.merge(right=df_geo_mun, on='id_municipio', how='left')

# Results
gdf_mun.info()
gdf_mun.head(2)

<br>

## Dados Tabulares: Municípios


df_tjsp_mun = pd.read_csv(output_path_tab / 'Municipios.csv')

df_tjsp_mun = df_tjsp_mun.drop(
    ['municipio_tjsp', 'municipio_corrigido', 'id_cj'], axis=1, errors='ignore'
)

df_tjsp_mun.info()
df_tjsp_mun.head(2)

# Junta Delimitação com Nomes Corretos
gdf_mun = gdf_mun.merge(
    right=df_tjsp_mun,
    left_on='id_municipio',
    right_on='id_municipio',
    how='left',
)

# Delete id_municipio
gdf_mun = gdf_mun.drop(['id_municipio'], axis=1, errors='ignore')

# Results
gdf_mun.info()
gdf_mun.head(2)

<br>

## Dados Tabulares: Comarcas


df_comarca = pd.read_csv(output_path_tab / 'Comarcas.csv')

# Results
df_comarca.info()
df_comarca.head(2)

# Junta Delimitação com Nomes Corretos
gdf_mun = gdf_mun.merge(
    right=df_comarca, left_on='id_comarca', right_on='id_comarca', how='left'
)

# Results
gdf_mun.info()
gdf_mun.head(2)

<br>

## Dados Tabulares: CJs


df_cjs = pd.read_csv(output_path_tab / 'Circunscrições Judiciárias.csv')

# Results
df_cjs.info()
df_cjs.head(2)

# Junta Delimitação com CJs
gdf_mun = gdf_mun.merge(
    right=df_cjs, left_on='id_cj', right_on='id_cj', how='left'
)

# Results
gdf_mun.info()
gdf_mun.head(2)

<br>

## Dados Tabulares: RAJs


df_rajs = pd.read_csv(output_path_tab / 'RAJs.csv')

# Results
df_rajs.info()
df_rajs.head(2)

# Junta Delimitação com RAJs
gdf_mun = gdf_mun.merge(
    right=df_rajs, left_on='id_raj', right_on='id_raj', how='left'
)

# Results
gdf_mun.info()
gdf_mun.head(2)

gdf_mun.to_file(
    output_path_gpkg / 'sp_tjsp_divadmin.gpkg',
    driver='GPKG',
)

converts.files.convert_to_7zip(output_path_gpkg, output_path_geo)
